Add outline of extension
This updates the manifest.json to make this only run on concur and add a
few files we'll need later. Content.css will have the styles for the
icons to display. Content.js will find the address for each hotel,
standardize it and match according to the database. Popup.html will be
the popup for the chrome extension and data.json is a temporary file
until APIs comes in for testing.